requirements.txt
# Core dependencies
disnake>=2.9.1
python-dotenv>=1.0.0
PyYAML>=6.0.1
aiohttp>=3.9.1
requests>=2.31.0
asyncpg>=0.29.0

# Development tools
black>=24.1.0  # Code formatting
isort>=5.13.2  # Import sorting
flake8>=7.0.0  # Linting
mypy>=1.8.0   # Type checking
pytest>=8.0.0  # Testing
pytest-asyncio>=0.23.5  # Async test support
pytest-cov>=4.1.0      # Test coverage